# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PaperMatrix API
  version: 0.1.0
  description: Local-first file-backed research paper management API.
servers:
  - url: http://127.0.0.1:8000/api/v1
paths:
...
  /paper-sources/scan:
    post:
      operationId: scanPaperDirectory
      summary: 扫描允许目录中的 PDF 候选，不写入项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory: { type: string }
                recursive: { type: boolean, default: false }
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                type: object
                required: [scan_token, items]
                properties:
                  scan_token: { type: string }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PaperCandidate" }
                  warnings:
                    type: array
                    items: { type: string }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listPapers
      summary: 项目论文列表
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: source_status
          schema:
            {
              type: string,
              enum: [available, missing, changed, unreadable, unlinked],
            }
        - in: query
          name: group
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: sort
          schema: { type: string, default: -updated_at }
      responses:
        "200":
          description: Paper list
          content:
            application/json:
              schema:
                type: object
                required:
                  [items, invalid_items, total, invalid_total, page, page_size]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PaperSummary" }
                  invalid_items:
                    type: array
                    items: { $ref: "#/components/schemas/InvalidPaperRecord" }
                  total: { type: integer }
                  invalid_total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
  /projects/{project_id}/papers/import:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: importPapers
      summary: 登记扫描得到的 PDF 引用，不复制文件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scan_token, candidate_ids]
              properties:
                scan_token: { type: string }
                candidate_ids:
                  type: array
                  minItems: 1
                  items: { type: string }
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /projects/{project_id}/papers/upload:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: uploadPaperForInspection
      summary: 瞬时读取单个 PDF 并创建未关联记录，不保存 PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: [string, "null"] }
      responses:
        "201":
          description: Created unlinked paper record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
        "413": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/link:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: linkSinglePaper
      summary: 登记允许目录内的单个 PDF 路径
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string }
      responses:
        "201":
          description: Created linked paper record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
  /projects/{project_id}/papers/manual:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: createManualPaper
      summary: 创建尚未关联 PDF 的论文记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        "201":
          description: Created unlinked paper record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
  /projects/{project_id}/papers/{paper_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaper
      responses:
        "200":
          description: Paper
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    patch:
      operationId: updatePaper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperUpdate"
      responses:
        "200":
          description: Updated paper
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deletePaperRecord
      summary: 删除项目中的记录和笔记，不删除源 PDF
      parameters:
        - in: query
          name: confirm_metadata_only
          required: true
          schema: { type: boolean }
      responses:
        "200":
          description: Deleted PaperMatrix files
          content:
            application/json:
              schema:
                type: object
                required: [source_pdf_untouched, removed_files]
                properties:
                  source_pdf_untouched: { type: boolean, const: true }
                  removed_files:
                    type: array
                    items: { type: string }
  /projects/{project_id}/papers/{paper_id}/relink:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: relinkPaper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_path, expected_revision]
              properties:
                new_path: { type: string }
                expected_revision: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Relinked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Paper" }
...
  /projects/{project_id}/papers/{paper_id}/analysis:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaperAnalysis
      summary: 获取论文中可用于比较和可视化的结构化分析条目
      responses:
        "200":
          description: Paper analysis document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperAnalysisDocument" }
...
  /projects/{project_id}/papers/{paper_id}/analysis/items:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: createPaperAnalysisItem
      summary: 添加带稳定 ID 和证据引用的分析条目
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalysisItemInput" }
      responses:
        "201":
          description: Updated analysis document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/analysis/items/{item_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
      - in: path
        name: item_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updatePaperAnalysisItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalysisItemInput" }
      responses:
        "200":
          description: Updated analysis document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deletePaperAnalysisItem
      parameters:
        - in: query
          name: expected_revision
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Updated analysis document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  parameters:
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
    PaperId:
      in: path
      name: paper_id
      required: true
      schema: { type: string, format: uuid }
  responses:
    ErrorResponse:
      description: PaperMatrix error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            action: { type: [string, "null"] }
            request_id: { type: string }
...
    PaperCandidate:
      type: object
      required:
        [candidate_id, display_path, filename, title, size_bytes, readable]
      properties:
        candidate_id: { type: string }
        display_path: { type: string }
        filename: { type: string }
        title: { type: string }
        size_bytes: { type: integer, minimum: 0 }
        page_count: { type: [integer, "null"] }
        readable: { type: boolean }
    ImportResult:
      type: object
      required: [imported, skipped]
      properties:
        imported:
          type: array
          items: { $ref: "#/components/schemas/Paper" }
        skipped:
          type: array
          items:
            type: object
            properties:
              candidate_id: { type: string }
              reason: { type: string }
    PaperSummary:
      type: object
      required:
        [
          paper_id,
          project_id,
          title,
          authors,
          affiliations,
          topics,
          tags,
          keywords,
          source_status,
          reading_status,
          updated_at,
          revision,
        ]
      properties:
        paper_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        title: { type: string }
        short_title: { type: string }
        authors:
          type: array
          items: { type: string }
        affiliations:
          type: array
          items: { type: string }
        year: { type: [integer, "null"] }
        venue: { type: [string, "null"] }
        publication_date: { type: [string, "null"], format: date }
        reading_date: { type: [string, "null"], format: date }
        citation_count: { type: [integer, "null"], minimum: 0 }
        language: { type: [string, "null"] }
        keywords:
          type: array
          items: { type: string }
        group: { type: [string, "null"] }
        topics:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        reading_status:
          {
            type: string,
            enum: [unread, skimmed, deep_read, summarized, reported],
          }
        importance_score: { type: [integer, "null"], minimum: 1, maximum: 5 }
        writing_uses:
          type: array
          items: { type: string }
        source_status:
          {
            type: string,
            enum: [available, missing, changed, unreadable, unlinked],
          }
        source_filename: { type: [string, "null"] }
        page_count: { type: [integer, "null"], minimum: 1 }
        one_sentence_summary: { type: string }
        updated_at: { type: string, format: date-time }
        revision: { type: integer, minimum: 1 }
    InvalidPaperRecord:
      type: object
      required: [paper_id, title, schema_version, reason]
      properties:
        paper_id: { type: string, format: uuid }
        title: { type: string }
        schema_version: { type: [integer, "null"] }
        reason: { type: string }
    Paper:
      type: object
      required:
        [
          schema_version,
          paper_id,
          project_id,
          source,
          bibliography,
          organization,
          structured_summary,
          created_at,
          updated_at,
          revision,
        ]
      properties:
        schema_version: { type: integer, const: 11 }
        paper_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        source:
          type: object
          required:
            [
              path,
              path_mode,
              original_filename,
              size_bytes,
              modified_at,
              fingerprint,
              sha256,
              page_count,
              status,
            ]
          properties:
            path: { type: [string, "null"] }
            path_mode:
              {
                type: [string, "null"],
                enum: [absolute, workspace_relative, null],
              }
            original_filename: { type: [string, "null"] }
            size_bytes: { type: [integer, "null"] }
            modified_at: { type: [string, "null"], format: date-time }
            fingerprint: { type: [string, "null"] }
            sha256: { type: [string, "null"] }
            page_count: { type: [integer, "null"] }
            status:
              {
                type: string,
                enum: [available, missing, changed, unreadable, unlinked],
              }
        bibliography: { type: object, additionalProperties: true }
        organization: { type: object, additionalProperties: true }
        structured_summary:
          type: object
          additionalProperties: true
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/AnalysisItem" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        revision: { type: integer, minimum: 1 }
    PaperUpdate:
      type: object
      required:
        [
          title,
          authors,
          affiliations,
          keywords,
          abstract_text,
          expected_revision,
        ]
      properties:
        expected_revision: { type: integer, minimum: 1 }
        title: { type: string }
        short_title: { type: string }
        authors:
          type: array
          items: { type: string }
        affiliations:
          type: array
          items: { type: string }
        venue: { type: [string, "null"] }
        publication_date: { type: [string, "null"], format: date }
        reading_date: { type: [string, "null"], format: date }
        citation_count: { type: [integer, "null"], minimum: 0 }
        language: { type: [string, "null"] }
        keywords:
          type: array
          items: { type: string }
        abstract_text: { type: string }
        urls:
          type: array
          items: { type: string, format: uri }
        code_url: { type: [string, "null"], format: uri }
        data_url: { type: [string, "null"], format: uri }
        group: { type: [string, "null"] }
        reading_status:
          type: string
          enum: [unread, skimmed, deep_read, summarized, reported]
        importance_score: { type: [integer, "null"], minimum: 1, maximum: 5 }
        one_sentence_summary: { type: string, maxLength: 5000 }
...
    EvidenceReference:
      type: object
      required:
        [
          paper_id,
          evidence_code,
          page_label,
          pdf_page_index,
          section,
          figure,
          table,
          locator_note,
        ]
      properties:
        evidence_id: { type: string, format: uuid }
        evidence_code: { type: [string, "null"], maxLength: 40 }
        paper_id: { type: string, format: uuid }
        page_label: { type: [string, "null"] }
        pdf_page_index: { type: [integer, "null"], minimum: 1 }
        section: { type: [string, "null"] }
        figure: { type: [string, "null"] }
        table: { type: [string, "null"] }
        locator_note: { type: string }
...
    AnalysisItem:
      type: object
      required:
        [
          item_id,
          kind,
          display_label,
          title,
          summary,
          section_key,
          section_title,
          section_order,
          source_anchor,
          source_note_revision,
          source_fingerprint,
          attributes,
          evidence_ids,
          tags,
          writing_uses,
          is_favorite,
          created_at,
          updated_at,
        ]
      properties:
        item_id: { type: string, format: uuid }
        kind:
          type: string
          enum:
            [
              background,
              research_problem,
              scenario,
              related_work,
              method,
              method_component,
              mechanism,
              challenge,
              innovation,
              contribution,
              experiment,
              finding,
              author_limitation,
              reviewer_limitation,
              condition,
            ]
        display_label: { type: [string, "null"], maxLength: 80 }
        title: { type: string, minLength: 1, maxLength: 300 }
        summary: { type: string, maxLength: 20000 }
        section_key: { type: [string, "null"], maxLength: 40 }
        section_title: { type: [string, "null"], maxLength: 300 }
        section_order: { type: [integer, "null"], minimum: 1 }
        source_anchor: { type: [string, "null"], maxLength: 80 }
        source_note_revision: { type: [integer, "null"], minimum: 1 }
        source_fingerprint:
          type: [string, "null"]
          pattern: "^[0-9a-f]{64}$"
        attributes:
          type: object
          additionalProperties: { type: string }
        evidence_ids:
          type: array
          items: { type: string, format: uuid }
          uniqueItems: true
        tags:
          type: array
          items: { type: string }
          uniqueItems: true
        writing_uses:
          type: array
          items:
            type: string
            enum:
              [
                INTRO,
                RELATED,
                METHOD,
                BASELINE,
                DATASET,
                METRIC,
                LIMITATION,
                DISCUSSION,
                FUTURE,
              ]
          uniqueItems: true
        is_favorite: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AnalysisItemInput:
      type: object
      required:
        [
          kind,
          title,
          summary,
          attributes,
          tags,
          writing_uses,
          expected_revision,
        ]
      properties:
        kind:
          type: string
          enum:
            [
              background,
              research_problem,
              scenario,
              related_work,
              method,
              method_component,
              mechanism,
              challenge,
              innovation,
              contribution,
              experiment,
              finding,
              author_limitation,
              reviewer_limitation,
              condition,
            ]
        display_label: { type: [string, "null"], maxLength: 80 }
        title: { type: string, minLength: 1, maxLength: 300 }
        summary: { type: string, maxLength: 20000 }
        attributes:
          type: object
          additionalProperties: { type: string }
        tags:
          type: array
          items: { type: string }
        writing_uses:
          type: array
          items:
            type: string
            enum:
              [
                INTRO,
                RELATED,
                METHOD,
                BASELINE,
                DATASET,
                METRIC,
                LIMITATION,
                DISCUSSION,
                FUTURE,
              ]
        expected_revision: { type: integer, minimum: 1 }
    PaperAnalysisDocument:
      type: object
      required: [paper_id, revision, updated_at, evidence_catalog, items]
      properties:
        paper_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 1 }
        updated_at: { type: string, format: date-time }
        evidence_catalog:
          type: array
          items: { $ref: "#/components/schemas/EvidenceReference" }
        items:
          type: array
          items: { $ref: "#/components/schemas/AnalysisItem" }
...